Modified the code to use UInt32 instead of UInt16

+++ Sources/PDFjet/FlateLiteral.swift
@@ -23,9 +23,7 @@ SOFTWARE.
 */
 import Foundation
 
-
 public class FlateLiteral {
-
     //  Huffman codes for the literal alphabet:
     //  ==========================================
     //  Literal      nBits       Codes
@@ -34,29 +32,26 @@ public class FlateLiteral {
     //                           10111111
     //  144 - 255     9          110010000 through
     //                           111111111
-
     static let instance = FlateLiteral()
 
-    var codes = [UInt16]()
+    var codes = [UInt32]()
     var nBits = [UInt8]()
 
     private init() {
-        var code: UInt16 = 0b00110000
+        var code: UInt32 = 0b00110000
         var i = 0
         while i < 144 {
-            codes.append(UInt16(FlateUtils.reverse(UInt32(code), length: 8)))
+            codes.append(UInt32(FlateUtils.reverse(UInt32(code), length: 8)))
             nBits.append(UInt8(8))
             code += 1
             i += 1
         }
         code = 0b110010000
         while i < 256 {
-            codes.append(UInt16(FlateUtils.reverse(UInt32(code), length: 9)))
+            codes.append(UInt32(FlateUtils.reverse(UInt32(code), length: 9)))
             nBits.append(UInt8(9))
             code += 1
             i += 1
         }
     }
-
 }
-
